+++ nessi/globopt/pso/swarm.py
@@ -8,8 +8,8 @@
 # Copyright (C) 2018 Damien Pageot
 # ------------------------------------------------------------------
 """
-The Swarm class and its associated functions for particles swarm
-optimization.
+Class and methods for particle swarm optimization.
+
 :copyright:
     Damien Pageot (nessi.develop@protonmail.com)
 :license:
@@ -17,8 +17,11 @@ optimization.
     (https://www.gnu.org/copyleft/lesser.html)
 """
 
-import numpy as np
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
 
+import numpy as np
 
 class Swarm():
     """
@@ -135,7 +138,7 @@ class Swarm():
         if topology == 'ring':
             ibest = indv
             vbest = self.misfit[indv]
-            for i in range(indv-1, indv+1):
+            for i in range(indv-1, indv+2):
                 ii = i
                 if i < 0:
                     ii = nindv-1
@@ -162,7 +165,7 @@ class Swarm():
         """
         Standard PSO update.
 
-        :param control: 0 for weight (default), 1 for constriction 
+        :param control: 0 for weight (default), 1 for constriction
         :param c_0: value of the control parameter (default 0.7298)
         :param c_1: value of the cognitive parameter (default 2.05)
         :param c_2: value of the social parameter (default 2.05)
@@ -202,7 +205,6 @@ class Swarm():
                                                 * (gbest[ipts, ipar]-current)
 
                     # Check particle velocity
-                    # vvv = particles[ibest]['history'][ipts,ipar]
                     if(np.abs(self.velocity[indv, ipts, ipar]) > self.pspace[ipts, ipar, 2]):
                         self.velocity[indv, ipts, ipar] = \
                             np.sign(self.velocity[indv, ipts, ipar])\