+++ nessi/modbuilder/__init__.py
@@ -1,13 +1,13 @@
 # -*- coding: utf-8 -*-
 # -------------------------------------------------------------------
-# Filename: Convenience import for nessi.modeling.interfaces
+# Filename: Convenience import for nessi.modbuilder
 #   Author: Damien Pageot
 #    Email: nessi.develop@protonmail.com
 #
 # Copyright (C) 2018 Damien Pageot
 # ------------------------------------------------------------------
 """
-Initialization file for nessi.modeling.interfaces .
+Initialization file for nessi.modbuilder .
 
 :copyright:
     Damien Pageot (nessi.develop@protonmail.com)
@@ -20,10 +20,11 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-# Import nessi.modeling.interfaces classes and functions
-from .gpdcwrap import dispersion_curve_init
-from .gpdcwrap import dispersion_curve_rayleigh
-from .gpdcwrap import dispersion_curve_love
+# Import nessi.modbuilder.interp2d classes and functions
+from interp2d.grdwrap import voronoi
+from interp2d.grdwrap import idweight
+from interp2d.grdwrap import sibson1
+from interp2d.grdwrap import sibson2
 
 if __name__ == '__main__':
     import doctest