+++ nessi/modeling/__init__.py
@@ -1,15 +1,26 @@
 # -*- coding: utf-8 -*-
+# -------------------------------------------------------------------
+# Filename: Convenience import for nessi.grd
+#   Author: Damien Pageot
+#    Email: nessi.develop@protonmail.com
+#
+# Copyright (C) 2018 Damien Pageot
+# ------------------------------------------------------------------
 """
-nessi.swm
+Initialization file for nessi.grd .
 
-nessi.interfaces
-
-================================================
+:copyright:
+    Damien Pageot (nessi.develop@protonmail.com)
+:license:
+    GNU Lesser General Public License, Version 3
+    (https://www.gnu.org/copyleft/lesser.html)
 """
-from __future__ import (absolute_import,
-                        division,
-                        print_function,
-                        unicode_literals)
+
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+# Import nessi.grd classes and functions
 
 
 if __name__ == '__main__':